refactor: Split item update planning
